Added optional --output parameter to jig
Added optional --output parameter for the prepare script. The user can specify in the --opts arg a path where to mount the host folder specified in --output to the "/output" guest path in order to save the index or other data created during the indexing/initialization process. If not provided no host directory is mounted in the "/output" guest path.